Added Early Stuff for modeling etc
Playing around with a bunch of different ways to look at the stripe
formation and compare it to the gap gene concentrations.  Another
crucial change, AP positions now correspond to the nucleus position of
the corresponding particle, this should fix fractions of active nuclei
that would come up greater than 1.  ~~ Also analyze.m is a script w/
hardcoded file paths to run most of the stuff I have now on the data.
makeFigures.m takes the workspace generated by analyze and makes/saves
a bunch of pretty figures.